Panda preprocessing expression
In Panda preprocessing there was a problem with indices. Using gene2idx.get(x, 0) always give you the index 0 if x is missing from gene2idx.get (like a gene in gene expression and not in motif, since gene2idx is build on top of the intersection of expression and motif). Now we use gene_names to both create the indices for self.expression and to access with .loc[] the expression data frame self.expression_data